+++ frontend/src/scenes/todaysAgenda.jsx
@@ -1,12 +1,12 @@
 import { Box, Container, Typography } from "@mui/material";
 import Body from "../components/containers/body";
-import WeeklyCalendar from "../components/calendar/weeklyCalendar";
+import CalendarView from "../components/calendar/calendarView";
 
 function TodaysAgenda(){
 
     return(
         <Body>
-            <WeeklyCalendar/>
+            <CalendarView/>
         </Body>
     )
 }
